+++ xarray/backends/rasterio_.py
@@ -301,6 +301,18 @@ def open_rasterio(filename, parse_coordinates=None, chunks=None, cache=None,
         attrs['nodatavals'] = tuple(
             np.nan if nodataval is None else nodataval
             for nodataval in riods.nodatavals)
+    if hasattr(riods, 'scales'):
+        # The scale values for the raster bands
+        attrs['scales'] = riods.scales
+    if hasattr(riods, 'offsets'):
+        # The offset values for the raster bands
+        attrs['offsets'] = riods.offsets
+    if hasattr(riods, 'descriptions') and any(riods.descriptions):
+        # Descriptions for each dataset band
+        attrs['descriptions'] = riods.descriptions
+    if hasattr(riods, 'units') and any(riods.units):
+        # A list of units string for each dataset band
+        attrs['units'] = riods.units
 
     # Parse extra metadata from tags, if supported
     parsers = {'ENVI': _parse_envi}